la mise en place de confirmation pour delete

+++ client/src/components/CardAdmin/TableArtwork/TableArtwork.tsx
@@ -10,20 +10,24 @@ function TableArtwork({ artworks }: { artworks: Oeuvre[] }) {
   const [artworksAdmin, setArtworksAdmin] = useState<Oeuvre[]>(artworks);
   const handleDelete = async (id: number) => {
     try {
-      // J'utilise ma route API pour delete mon oeuvre
-      api.delete(`/api/admin/${id}?q=oeuvres`);
+      if (confirm("Êtes-vous sûr de vouloir supprimer cet tabeau ?")) {
+        // J'utilise ma route API pour delete mon oeuvre
+        api.delete(`/api/admin/${id}?q=oeuvres`);
 
-      /**
-       * Ensuite, grâce à la méthode filter, je crée un nouveau tableau
-       * sans mon oeuvre supprimée.
-       */
-      const newArtworks: Oeuvre[] = artworksAdmin.filter(
-        (artwork: Oeuvre) => artwork.id !== id,
-      );
+        /**
+         * Ensuite, grâce à la méthode filter, je crée un nouveau tableau
+         * sans mon oeuvre supprimée.
+         */
+        const newArtworks: Oeuvre[] = artworksAdmin.filter(
+          (artwork: Oeuvre) => artwork.id !== id,
+        );
 
-      // Je mets à jour mon state avec mon nouveau tableau
-      setArtworksAdmin(newArtworks);
-      successToast("L'oeuvre est bien supprimé");
+        // Je mets à jour mon state avec mon nouveau tableau
+        setArtworksAdmin(newArtworks);
+        successToast("L'oeuvre est bien supprimé");
+      } else {
+        alert("Action annulée.");
+      }
     } catch (error) {
       failureToast("Oups, une erreur est survenu");
       console.error(error);
